+++ lib/chainsync/chainsync/db/hyperdrive/db_interface.test.py
@@ -1,0 +1,377 @@
+"""CRUD tests for Transaction"""
+from datetime import datetime
+from decimal import Decimal
+
+import numpy as np
+import pytest
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
+
+from ..base.db_interface import get_latest_block_number_from_table
+from ..base.db_schema import Base
+from .db_interface import (
+    add_checkpoint_infos,
+    add_pool_config,
+    add_pool_infos,
+    add_transactions,
+    add_wallet_deltas,
+    add_wallet_infos,
+    get_agents,
+    get_all_wallet_info,
+    get_checkpoint_info,
+    get_current_wallet_info,
+    get_latest_block_number_from_pool_info_table,
+    get_latest_block_number_from_table,
+    get_pool_config,
+    get_pool_info,
+    get_transactions,
+    get_wallet_deltas,
+)
+from .db_schema import CheckpointInfo, HyperdriveTransaction, PoolConfig, PoolInfo, WalletDelta, WalletInfo
+
+engine = create_engine("sqlite:///:memory:")  # in-memory SQLite database for testing
+Session = sessionmaker(bind=engine)
+
+# fixture arguments in test function have to be the same as the fixture name
+# pylint: disable=redefined-outer-name
+
+
+@pytest.fixture(scope="function")
+def session():
+    """Session fixture for tests"""
+    Base.metadata.create_all(engine)  # create tables
+    session_ = Session()
+    yield session_
+    session_.close()
+    Base.metadata.drop_all(engine)  # drop tables
+
+
+class TestTransactionInterface:
+    """Testing postgres interface for transaction table"""
+
+    def test_latest_block_number(self, session):
+        """Testing retrevial of transaction via interface"""
+        transaction_1 = HyperdriveTransaction(blockNumber=1, transactionHash="a", event_value=Decimal("3.0"))
+        add_transactions([transaction_1], session)
+
+        latest_block_number = get_latest_block_number_from_table(HyperdriveTransaction, session)
+        assert latest_block_number == 1
+
+        transaction_2 = HyperdriveTransaction(blockNumber=2, transactionHash="b", event_value=Decimal("3.2"))
+        transaction_3 = HyperdriveTransaction(blockNumber=3, transactionHash="c", event_value=Decimal("3.4"))
+        add_transactions([transaction_2, transaction_3], session)
+
+        latest_block_number = get_latest_block_number_from_table(HyperdriveTransaction, session)
+        assert latest_block_number == 3
+
+    def test_get_transactions(self, session):
+        """Testing retrevial of transactions via interface"""
+        transaction_1 = HyperdriveTransaction(blockNumber=0, transactionHash="a", event_value=Decimal("3.1"))
+        transaction_2 = HyperdriveTransaction(blockNumber=1, transactionHash="b", event_value=Decimal("3.2"))
+        transaction_3 = HyperdriveTransaction(blockNumber=2, transactionHash="c", event_value=Decimal("3.3"))
+        add_transactions([transaction_1, transaction_2, transaction_3], session)
+
+        transactions_df = get_transactions(session)
+        np.testing.assert_array_equal(transactions_df["event_value"], [3.1, 3.2, 3.3])
+
+    def test_block_query_transactions(self, session):
+        """Testing querying by block number of transactions via interface"""
+        transaction_1 = HyperdriveTransaction(blockNumber=0, transactionHash="a", event_value=Decimal("3.1"))
+        transaction_2 = HyperdriveTransaction(blockNumber=1, transactionHash="b", event_value=Decimal("3.2"))
+        transaction_3 = HyperdriveTransaction(blockNumber=2, transactionHash="c", event_value=Decimal("3.3"))
+        add_transactions([transaction_1, transaction_2, transaction_3], session)
+
+        transactions_df = get_transactions(session, start_block=1)
+        np.testing.assert_array_equal(transactions_df["event_value"], [3.2, 3.3])
+
+        transactions_df = get_transactions(session, start_block=-1)
+        np.testing.assert_array_equal(transactions_df["event_value"], [3.3])
+
+        transactions_df = get_transactions(session, end_block=1)
+        np.testing.assert_array_equal(transactions_df["event_value"], [3.1])
+
+        transactions_df = get_transactions(session, end_block=-1)
+        np.testing.assert_array_equal(transactions_df["event_value"], [3.1, 3.2])
+
+        transactions_df = get_transactions(session, start_block=1, end_block=-1)
+        np.testing.assert_array_equal(transactions_df["event_value"], [3.2])
+
+
+class TestCheckpointInterface:
+    """Testing postgres interface for checkpoint table"""
+
+    def test_latest_block_number(self, session):
+        """Testing retrevial of checkpoint via interface"""
+        checkpoint_1 = CheckpointInfo(blockNumber=1, timestamp=datetime.now())
+        add_checkpoint_infos([checkpoint_1], session)
+        session.commit()
+
+        latest_block_number = get_latest_block_number_from_table(CheckpointInfo, session)
+        assert latest_block_number == 1
+
+        checkpoint_2 = CheckpointInfo(blockNumber=2, timestamp=datetime.now())
+        checkpoint_3 = CheckpointInfo(blockNumber=3, timestamp=datetime.now())
+        add_checkpoint_infos([checkpoint_2, checkpoint_3], session)
+
+        latest_block_number = get_latest_block_number_from_table(CheckpointInfo, session)
+        assert latest_block_number == 3
+
+    def test_get_checkpoints(self, session):
+        """Testing retrevial of checkpoints via interface"""
+        date_1 = datetime(1945, 8, 6)
+        date_2 = datetime(1984, 8, 9)
+        date_3 = datetime(2001, 9, 11)
+        checkpoint_1 = CheckpointInfo(blockNumber=0, timestamp=date_1)
+        checkpoint_2 = CheckpointInfo(blockNumber=1, timestamp=date_2)
+        checkpoint_3 = CheckpointInfo(blockNumber=2, timestamp=date_3)
+        add_checkpoint_infos([checkpoint_1, checkpoint_2, checkpoint_3], session)
+
+        checkpoints_df = get_checkpoint_info(session)
+        np.testing.assert_array_equal(
+            checkpoints_df["timestamp"].dt.to_pydatetime(), np.array([date_1, date_2, date_3])
+        )
+
+    def test_block_query_checkpoints(self, session):
+        """Testing querying by block number of checkpoints via interface"""
+        checkpoint_1 = CheckpointInfo(blockNumber=0, timestamp=datetime.now(), sharePrice=Decimal("3.1"))
+        checkpoint_2 = CheckpointInfo(blockNumber=1, timestamp=datetime.now(), sharePrice=Decimal("3.2"))
+        checkpoint_3 = CheckpointInfo(blockNumber=2, timestamp=datetime.now(), sharePrice=Decimal("3.3"))
+        add_checkpoint_infos([checkpoint_1, checkpoint_2, checkpoint_3], session)
+
+        checkpoints_df = get_checkpoint_info(session, start_block=1)
+        np.testing.assert_array_equal(checkpoints_df["sharePrice"], [3.2, 3.3])
+
+        checkpoints_df = get_checkpoint_info(session, start_block=-1)
+        np.testing.assert_array_equal(checkpoints_df["sharePrice"], [3.3])
+
+        checkpoints_df = get_checkpoint_info(session, end_block=1)
+        np.testing.assert_array_equal(checkpoints_df["sharePrice"], [3.1])
+
+        checkpoints_df = get_checkpoint_info(session, end_block=-1)
+        np.testing.assert_array_equal(checkpoints_df["sharePrice"], [3.1, 3.2])
+
+        checkpoints_df = get_checkpoint_info(session, start_block=1, end_block=-1)
+        np.testing.assert_array_equal(checkpoints_df["sharePrice"], [3.2])
+
+
+class TestPoolConfigInterface:
+    """Testing postgres interface for poolconfig table"""
+
+    def test_get_pool_config(self, session):
+        """Testing retrevial of pool config via interface"""
+        pool_config_1 = PoolConfig(contractAddress="0", initialSharePrice=Decimal("3.2"))
+        add_pool_config(pool_config_1, session)
+
+        pool_config_df_1 = get_pool_config(session, coerce_float=False)
+        assert len(pool_config_df_1) == 1
+        assert pool_config_df_1.loc[0, "initialSharePrice"] == Decimal("3.2")
+
+        pool_config_2 = PoolConfig(contractAddress="1", initialSharePrice=Decimal("3.4"))
+        add_pool_config(pool_config_2, session)
+
+        pool_config_df_2 = get_pool_config(session, coerce_float=False)
+        assert len(pool_config_df_2) == 2
+        np.testing.assert_array_equal(pool_config_df_2["initialSharePrice"], np.array([Decimal("3.2"), Decimal("3.4")]))
+
+    def test_primary_id_query_pool_config(self, session):
+        """Testing retrevial of pool config via interface"""
+        pool_config = PoolConfig(contractAddress="0", initialSharePrice=Decimal("3.2"))
+        add_pool_config(pool_config, session)
+
+        pool_config_df_1 = get_pool_config(session, contract_address="0", coerce_float=False)
+        assert len(pool_config_df_1) == 1
+        assert pool_config_df_1.loc[0, "initialSharePrice"] == Decimal("3.2")
+
+        pool_config_df_2 = get_pool_config(session, contract_address="1", coerce_float=False)
+        assert len(pool_config_df_2) == 0
+
+    def test_pool_config_verify(self, session):
+        """Testing retrevial of pool config via interface"""
+        pool_config_1 = PoolConfig(contractAddress="0", initialSharePrice=Decimal("3.2"))
+        add_pool_config(pool_config_1, session)
+        pool_config_df_1 = get_pool_config(session, coerce_float=False)
+        assert len(pool_config_df_1) == 1
+        assert pool_config_df_1.loc[0, "initialSharePrice"] == Decimal("3.2")
+
+        # Nothing should happen if we give the same pool_config
+        pool_config_2 = PoolConfig(contractAddress="0", initialSharePrice=Decimal("3.2"))
+        add_pool_config(pool_config_2, session)
+        pool_config_df_2 = get_pool_config(session, coerce_float=False)
+        assert len(pool_config_df_2) == 1
+        assert pool_config_df_2.loc[0, "initialSharePrice"] == Decimal("3.2")
+
+        # If we try to add another pool config with a different value, should throw a ValueError
+        pool_config_3 = PoolConfig(contractAddress="0", initialSharePrice=Decimal("3.4"))
+        with pytest.raises(ValueError):
+            add_pool_config(pool_config_3, session)
+
+
+class TestPoolInfoInterface:
+    """Testing postgres interface for poolinfo table"""
+
+    def test_latest_block_number(self, session):
+        """Testing latest block number call"""
+        timestamp_1 = datetime.fromtimestamp(1628472000)
+        pool_info_1 = PoolInfo(blockNumber=1, timestamp=timestamp_1)
+        add_pool_infos([pool_info_1], session)
+
+        latest_block_number = get_latest_block_number_from_pool_info_table(session)
+        assert latest_block_number == 1
+
+        timestamp_1 = datetime.fromtimestamp(1628472002)
+        pool_info_1 = PoolInfo(blockNumber=2, timestamp=timestamp_1)
+        timestamp_2 = datetime.fromtimestamp(1628472004)
+        pool_info_2 = PoolInfo(blockNumber=3, timestamp=timestamp_2)
+        add_pool_infos([pool_info_1, pool_info_2], session)
+
+        latest_block_number = get_latest_block_number_from_pool_info_table(session)
+        assert latest_block_number == 3
+
+    def test_get_pool_info(self, session):
+        """Testing retrevial of pool info via interface"""
+        timestamp_1 = datetime.fromtimestamp(1628472000)
+        pool_info_1 = PoolInfo(blockNumber=0, timestamp=timestamp_1)
+        timestamp_2 = datetime.fromtimestamp(1628472002)
+        pool_info_2 = PoolInfo(blockNumber=1, timestamp=timestamp_2)
+        timestamp_3 = datetime.fromtimestamp(1628472004)
+        pool_info_3 = PoolInfo(blockNumber=2, timestamp=timestamp_3)
+        add_pool_infos([pool_info_1, pool_info_2, pool_info_3], session)
+
+        pool_info_df = get_pool_info(session)
+        np.testing.assert_array_equal(
+            pool_info_df["timestamp"].dt.to_pydatetime(), np.array([timestamp_1, timestamp_2, timestamp_3])
+        )
+
+    def test_block_query_pool_info(self, session):
+        """Testing retrevial of pool info via interface"""
+        timestamp_1 = datetime.fromtimestamp(1628472000)
+        pool_info_1 = PoolInfo(blockNumber=0, timestamp=timestamp_1)
+        timestamp_2 = datetime.fromtimestamp(1628472002)
+        pool_info_2 = PoolInfo(blockNumber=1, timestamp=timestamp_2)
+        timestamp_3 = datetime.fromtimestamp(1628472004)
+        pool_info_3 = PoolInfo(blockNumber=2, timestamp=timestamp_3)
+        add_pool_infos([pool_info_1, pool_info_2, pool_info_3], session)
+        pool_info_df = get_pool_info(session, start_block=1)
+        np.testing.assert_array_equal(
+            pool_info_df["timestamp"].dt.to_pydatetime(), np.array([timestamp_2, timestamp_3])
+        )
+        pool_info_df = get_pool_info(session, start_block=-1)
+        np.testing.assert_array_equal(pool_info_df["timestamp"].dt.to_pydatetime(), np.array([timestamp_3]))
+        pool_info_df = get_pool_info(session, end_block=1)
+        np.testing.assert_array_equal(pool_info_df["timestamp"].dt.to_pydatetime(), np.array([timestamp_1]))
+        pool_info_df = get_pool_info(session, end_block=-1)
+        np.testing.assert_array_equal(
+            pool_info_df["timestamp"].dt.to_pydatetime(), np.array([timestamp_1, timestamp_2])
+        )
+        pool_info_df = get_pool_info(session, start_block=1, end_block=-1)
+        np.testing.assert_array_equal(pool_info_df["timestamp"].dt.to_pydatetime(), np.array([timestamp_2]))
+
+
+class TestWalletDeltaInterface:
+    """Testing postgres interface for walletinfo table"""
+
+    def test_latest_block_number(self, session):
+        """Testing retrevial of wallet info via interface"""
+        wallet_delta_1 = WalletDelta(blockNumber=1, transactionHash="a", delta=Decimal("3.0"))
+        add_wallet_deltas([wallet_delta_1], session)
+        latest_block_number = get_latest_block_number_from_table(WalletDelta, session)
+        assert latest_block_number == 1
+        wallet_delta_2 = WalletDelta(blockNumber=2, transactionHash="a", delta=Decimal("3.2"))
+        wallet_delta_3 = WalletDelta(blockNumber=3, transactionHash="a", delta=Decimal("3.4"))
+        add_wallet_deltas([wallet_delta_2, wallet_delta_3], session)
+        latest_block_number = get_latest_block_number_from_table(WalletDelta, session)
+        assert latest_block_number == 3
+
+    def test_get_wallet_delta(self, session):
+        """Testing retrevial of walletinfo via interface"""
+        wallet_delta_1 = WalletDelta(blockNumber=0, transactionHash="a", delta=Decimal("3.1"))
+        wallet_delta_2 = WalletDelta(blockNumber=1, transactionHash="a", delta=Decimal("3.2"))
+        wallet_delta_3 = WalletDelta(blockNumber=2, transactionHash="a", delta=Decimal("3.3"))
+        add_wallet_deltas([wallet_delta_1, wallet_delta_2, wallet_delta_3], session)
+        wallet_delta_df = get_wallet_deltas(session)
+        np.testing.assert_array_equal(wallet_delta_df["delta"], np.array([3.1, 3.2, 3.3]))
+
+    def test_block_query_wallet_delta(self, session):
+        """Testing querying by block number of wallet info via interface"""
+        wallet_delta_1 = WalletDelta(blockNumber=0, transactionHash="a", delta=Decimal("3.1"))
+        wallet_delta_2 = WalletDelta(blockNumber=1, transactionHash="a", delta=Decimal("3.2"))
+        wallet_delta_3 = WalletDelta(blockNumber=2, transactionHash="a", delta=Decimal("3.3"))
+        add_wallet_deltas([wallet_delta_1, wallet_delta_2, wallet_delta_3], session)
+        wallet_delta_df = get_wallet_deltas(session, start_block=1)
+        np.testing.assert_array_equal(wallet_delta_df["delta"], np.array([3.2, 3.3]))
+        wallet_delta_df = get_wallet_deltas(session, start_block=-1)
+        np.testing.assert_array_equal(wallet_delta_df["delta"], np.array([3.3]))
+        wallet_delta_df = get_wallet_deltas(session, end_block=1)
+        np.testing.assert_array_equal(wallet_delta_df["delta"], np.array([3.1]))
+        wallet_delta_df = get_wallet_deltas(session, end_block=-1)
+        np.testing.assert_array_equal(wallet_delta_df["delta"], np.array([3.1, 3.2]))
+        wallet_delta_df = get_wallet_deltas(session, start_block=1, end_block=-1)
+        np.testing.assert_array_equal(wallet_delta_df["delta"], np.array([3.2]))
+
+
+class TestWalletInfoInterface:
+    """Testing postgres interface for walletinfo table"""
+
+    def test_latest_block_number(self, session):
+        """Testing retrevial of wallet info via interface"""
+        wallet_info_1 = WalletInfo(blockNumber=1, tokenValue=Decimal("3.0"))
+        add_wallet_infos([wallet_info_1], session)
+        latest_block_number = get_latest_block_number_from_table(WalletInfo, session)
+        assert latest_block_number == 1
+        wallet_info_2 = WalletInfo(blockNumber=2, tokenValue=Decimal("3.2"))
+        wallet_info_3 = WalletInfo(blockNumber=3, tokenValue=Decimal("3.4"))
+        add_wallet_infos([wallet_info_2, wallet_info_3], session)
+        latest_block_number = get_latest_block_number_from_table(WalletInfo, session)
+        assert latest_block_number == 3
+
+    def test_get_wallet_info(self, session):
+        """Testing retrevial of walletinfo via interface"""
+        wallet_info_1 = WalletInfo(blockNumber=0, tokenValue=Decimal("3.1"))
+        wallet_info_2 = WalletInfo(blockNumber=1, tokenValue=Decimal("3.2"))
+        wallet_info_3 = WalletInfo(blockNumber=2, tokenValue=Decimal("3.3"))
+        add_wallet_infos([wallet_info_1, wallet_info_2, wallet_info_3], session)
+        wallet_info_df = get_all_wallet_info(session)
+        np.testing.assert_array_equal(wallet_info_df["tokenValue"], np.array([3.1, 3.2, 3.3]))
+
+    def test_block_query_wallet_info(self, session):
+        """Testing querying by block number of wallet info via interface"""
+        wallet_info_1 = WalletInfo(blockNumber=0, tokenValue=Decimal("3.1"))
+        wallet_info_2 = WalletInfo(blockNumber=1, tokenValue=Decimal("3.2"))
+        wallet_info_3 = WalletInfo(blockNumber=2, tokenValue=Decimal("3.3"))
+        add_wallet_infos([wallet_info_1, wallet_info_2, wallet_info_3], session)
+        wallet_info_df = get_all_wallet_info(session, start_block=1)
+        np.testing.assert_array_equal(wallet_info_df["tokenValue"], np.array([3.2, 3.3]))
+        wallet_info_df = get_all_wallet_info(session, start_block=-1)
+        np.testing.assert_array_equal(wallet_info_df["tokenValue"], np.array([3.3]))
+        wallet_info_df = get_all_wallet_info(session, end_block=1)
+        np.testing.assert_array_equal(wallet_info_df["tokenValue"], np.array([3.1]))
+        wallet_info_df = get_all_wallet_info(session, end_block=-1)
+        np.testing.assert_array_equal(wallet_info_df["tokenValue"], np.array([3.1, 3.2]))
+        wallet_info_df = get_all_wallet_info(session, start_block=1, end_block=-1)
+        np.testing.assert_array_equal(wallet_info_df["tokenValue"], np.array([3.2]))
+
+    def test_current_wallet_info(self, session):
+        """Testing helper function to get current wallet values"""
+        wallet_info_1 = WalletInfo(blockNumber=0, walletAddress="addr", tokenType="BASE", tokenValue=Decimal("3.1"))
+        wallet_info_2 = WalletInfo(blockNumber=1, walletAddress="addr", tokenType="LP", tokenValue=Decimal("5.1"))
+        add_wallet_infos([wallet_info_1, wallet_info_2], session)
+        wallet_info_df = get_current_wallet_info(session).reset_index()
+        np.testing.assert_array_equal(wallet_info_df["tokenType"], ["BASE", "LP"])
+        np.testing.assert_array_equal(wallet_info_df["tokenValue"], [3.1, 5.1])
+        # E.g., block 2, wallet base tokens gets updated to 6.1
+        wallet_info_3 = WalletInfo(blockNumber=2, walletAddress="addr", tokenType="BASE", tokenValue=Decimal("6.1"))
+        add_wallet_infos([wallet_info_3], session)
+        wallet_info_df = get_current_wallet_info(session).reset_index()
+        np.testing.assert_array_equal(wallet_info_df["tokenType"], ["BASE", "LP"])
+        np.testing.assert_array_equal(wallet_info_df["tokenValue"], [6.1, 5.1])
+
+    def test_get_agents(self, session):
+        """Testing helper function to get current wallet values"""
+        wallet_info_1 = WalletInfo(blockNumber=0, walletAddress="addr_1")
+        wallet_info_2 = WalletInfo(blockNumber=1, walletAddress="addr_1")
+        wallet_info_3 = WalletInfo(blockNumber=2, walletAddress="addr_2")
+        add_wallet_infos([wallet_info_1, wallet_info_2, wallet_info_3], session)
+        agents = get_agents(session)
+        assert len(agents) == 2
+        assert "addr_1" in agents
+        assert "addr_2" in agents
